Extract test mocks to dedicated module
Made this move because I intend to break down "test_distgit" into several files.
